:recycle: Organizes Types and Documents
